Factor more linear solver testing code into helpers
