+++ tests/Unit/ParallelAlgorithms/LinearSolver/CMakeLists.txt
@@ -8,9 +8,6 @@ set(LIBRARY_SOURCES
   Test_Tags.cpp
   )
 
-add_subdirectory(ConjugateGradient)
-add_subdirectory(Gmres)
-
 add_test_library(
   ${LIBRARY}
   "ParallelAlgorithms/LinearSolver/"
@@ -22,3 +19,50 @@ add_dependencies(
   ${LIBRARY}
   module_ConstGlobalCache
   )
+
+function(add_linear_solver_algorithm_test TEST_NAME)
+  set(EXECUTABLE_NAME Test_${TEST_NAME})
+  set(TEST_IDENTIFIER Integration.LinearSolver.${TEST_NAME})
+
+  add_spectre_executable(
+    ${EXECUTABLE_NAME}
+    ${EXECUTABLE_NAME}.cpp
+    )
+
+  add_dependencies(
+    ${EXECUTABLE_NAME}
+    module_ConstGlobalCache
+    module_Main
+    )
+
+  target_link_libraries(
+    ${EXECUTABLE_NAME}
+    PUBLIC
+    # Link against Boost::program_options for now until we have proper
+    # dependency handling for header-only libs
+    Boost::program_options
+    DataStructures
+    ErrorHandling
+    IO
+    Informer
+    ParallelLinearSolver
+    )
+
+  add_dependencies(test-executables ${EXECUTABLE_NAME})
+
+  add_test(
+    NAME "\"${TEST_IDENTIFIER}\""
+    COMMAND ${CMAKE_BINARY_DIR}/bin/${EXECUTABLE_NAME} --input-file
+    ${CMAKE_CURRENT_SOURCE_DIR}/${EXECUTABLE_NAME}.yaml
+    )
+
+  set_tests_properties(
+    "\"${TEST_IDENTIFIER}\""
+    PROPERTIES
+    TIMEOUT 5
+    LABELS "integration"
+    ENVIRONMENT "ASAN_OPTIONS=detect_leaks=0")
+endfunction()
+
+add_subdirectory(ConjugateGradient)
+add_subdirectory(Gmres)
